Update point-of-instantiation visibility text
Now we always consider the normal visibility and then
additionally consider the point of instantiation.

Additionally removed the bit about an implicit use for calls
that specify the module explicitly, since the relevant scoping
is now covered by applying the normal visibility rules to
the generic function's declaration.

This relates to PRs #8079 and #10363.