Add a note about explicit array return types and omitting domains/element types
This was the only place where array return types are mentioned.  I think it
isn't necessary to add this information to places that don't talk about array
return types at all, since the behavior in this case seems "expected" to me, but
others are welcome to disagree.

Spot-checked the generated spec.  Since no example was added, I don't feel the
need to do further testing.