When compileline fails, only error out in Qthreads config step.
Our solution for making it more obvious when the compileline utility had
failed during Qthreads builds had a flaw: it would halt the make not
only for builds but indeed for any target the Makefile was asked to
"make", including clean and clobber.  Since the third-party ordering is
such that GASNet is cleaned/clobbered before Qthreads is, this halted
Qthreads clean/clobber prematurely.

Here, restructure so that when compileline fails, during make's first
pass over the Makefile when it's processing make syntax, we just leave a
token in a make variable.  Then later, during make's second pass over
the Makefile when it's processing rules and recipes, check that make
variable in the QThreads configuration recipe and if it's set, error
out at that point.