Rename return-intent/return-type to yield-intent/yield-type for iterators
The spec repeatedly says you can't return anything from an iterator, but the
syntax for the declaration statement refers to return types and return intents.
Update the syntax to add a "yield-intent" and "yield-type" and mention that the
are syntactically identical to the return intent and return type in procedures,
though they refer to what is yielded from the iterator.

I had to add a syntax definition for these new terms or LaTeX would complain.