Convert a few other shootouts to use managed classes
This converts other shootout performance tests over to use managed
classes.  In the case of Albert Sidelnik's nbody, we can get away with
borrowed classes only resulting in a code that should work with older
versions of the compiler (albeit by leaking).  In the other cases, I
switched from unmanaged to managed pointers to use the language more
as we think we'd like.