Fix valgrind failure in literal-indexing.chpl
In the recently-added proc chpl_buildStandInRTT(type arrType: [])
the domain `standinDomain` was getting auto-destroyed before
that function returned. This resulted in access-after-delete in:

    test/arrays/vass/literal-indexing.chpl

To fix that, added pragma "no auto destroy".

As a result, that domain started to leak in:

    test/arrays/ferguson/from-iterator/skyline.chpl
    test/arrays/ferguson/from-iterator/inner-array-resize.chpl

Aside: these are probably the only two non-future tests that build
arrays of arrays using the chpl__initCopy() overload implementing
the shape-less conversion from an _iteratorRecord to an array.
The reason these tests passed valgrind before is that nothing
accessed the prematurely-destroyed `standinDomain`. The runtime
counterpart of the iteratorToArrayElementType()-computed type
was dropped on the floor. Somehow it wasn't stored in the
generic instantiation of the _ddata type of the `data` local
variable. (Maybe that is normal for instantiations of class types.)

To plug this leak, I auto-destroy it manually in that chpl__initCopy.
Since currently we do not have a way to access runtime types from
Chapel code, I added a primitive to do that.

It also required storing the iteratorToArrayElementType()-computed type
in a `type` alias so it can be accessed later. As a side benefit,
I fixed the case where the iterator for the _iteratorRecord formal
happens to yield no elements at run time, which probably never
worked before.