Convert programs directory to use delete-free features
This was more challenging than I'd expected and resulted in some
surprises (though I'm not sure that if the implementation had done
differently there wouldn't have been other surprises).

E.g., at one point I was surprised that:

      var x = new owned();
      var y = x;

would make 'y' take ownership of 'x' rather than borrowing from it.
Though on reflection it makes sense that this is the case or else
'x' would arguably be borrowed as well.

It was also much more difficult to delete elements from the middle
of the linked list than I expected due to the fact that the 'next'
field carries the ownership and so copying that out into an owned
temp requires some juggling.