Fix LCALS enum issues after enums quit having integral values by default
In a few places that are not used in the default testing path, LCALS was
using enums in ways that needed updates.

It cast an integer to an enum type but the enum didn't explicitly list any
integral values.  Assign 1 to the first enum symbol.

It was comparing an enum to an integer with '>'.  Cast the enum to int first.