shape preservation in the spec
Add the specification of shape preservation

* Improve how promotion is presented.
In particular, move the details of promotion over arrays
from the Arrays chapter to Data Parallelism chapter.
Make arrays reference that subsection so that we are consistent
in defining promotion for:

 - iterators
 - ranges
 - domains
 - arrays

* State that for-, forall- and promoted expressions
are iterators yielding one element at a time.
They become arrays when used to initialize a variable.
Remove Sung's earlier trial paragraph

* Create a table defining the domain of such an array
based on what kind of thing causes promotion.
The table is where promotion is defined.

* Have for- and forall-expressions reference that table
by pointing to the promotion section. If the for/forall expression
has a filtering predicate, then the resulting array has a
1-based one-dimensional domain.

* Move array-promotion sub-section from "Returning Arrays" section
to "Array Arguments" section, as it is fone for domain-promotion.
(The diffs show that the "Returning Arrays" section moved.)

* While there, remove \chpl from \subsubsection
as Latex gets upset otherwise.