Update remainder of library tests to remove unnecessary extern definitions
The extern definitions in test/compflags/library/withMain are still necessary
due to the lack of a #include of the generated header file.  I've chosen not to
add the #include there - the header is already included on the compile line, so
we might as well continue to benefit from the coverage.