Pre-release prep. Chapel 1.18.0: tools/mason/MasonUtils.chpl
NOTE: This is not Chapel release 1.18.0. This change prepares the source on
master branch for the upcoming Chapel 1.18.0 release, currently scheduled
for 20 Sep 2018.

This commit tries to help Mason pass automated testing on master branch
between now and when the parent commit,
"Pre-release prep. Chapel 1.18.0: post-release version markings on master",
is reversed.

Ie, this commit should be regarded as temporary, like the parent commit.