Attempt to make version button work better at release time
The current version of the version button logic for the docs looks
at whether or not the version string in doc/rst/conf.py contains
the text (pre-release) or not, which makes things a bit dicey
right around release time, like now, when we start pulling that
string from the version text.

This change uses a version variable with the version button logic
itself to keep track of which version is the next release and
adds the (pre-release) text there so that it will always be
applied.  This should also reduce the maintenance burden within
doc/rst/conf.py slightly.