Pre-release prep. Chapel 1.18.0: bump version numbers on master from 1.18 to 1.19
NOTE: This is not Chapel release 1.18.0 or 1.19.0. This change prepares the source on
master branch for next development cycle AFTER the upcoming Chapel 1.18.0 release,
currently scheduled for 20 Sep 2018.

The release/1.18 branch has been created, now Bump most of the "1.18" version
numbers in source on master branch to "1.19", to set up for the next development
cycle.