Optimize blocking puts/gets and fetching network AMOs
Optimize post_fma_and_wait to yield less frequently. This is similar to
c743dcb06b, where we had post_fma_and_wait_amo yield every 64 iterations, but
here we increase how often we yield the longer we wait. With this we will yield
every 32, then 16, 8, 4, 2, and finally every iteration. The benefit to doing
this is that if we end having a longer transactions, we'll still revert to
yielding every iteration. In practice, this is unlikely since post_fma_and_wait
is only used for 4/8 byte amos and puts/gets under 4096 bytes.