Fix updateDatfiles.py:remove_data_in_range to remove all occurrences of a date
This function had only been used on compiler performance .dat files before, and
they only have a single date. For multi-trial configs it was wrong because it
was only removing the first end-date. Update the script to find the last index
of the end date instead of the first index