Switch study version of binary trees from owned to unmanaged
This changes my study version of binary trees from owned to unmanaged,
but frees in a distinct step at the end rather than freeing while
we're computing a tree's sum.  By comparing to the previous and
release versions, this should help us distinguish the overhead for
using 'owned' from the overhead for deleting in a separate step
(requiring a re-traversal of the tree).