Remove regexdna.chpl
For awhile, we were preserving this after regexredux was added
to the CLBG competition, but at this point it's been retired
from the site, so it makes sense to retire it from this
submitted directory that mirrors the site as well.  For now,
I'm leaving it in the release directory for historical purposes,
like other benchmarks that have been retired from the site.