Solve internal error for param loop over improper range
Param for loops were assuming that the lower bound, higher bound, and stride of
a param loop would always be of type int, uint, or bool.  Additionally, it
assumed that they would always match, so that if the lower bound was one of the
proper types, the other two would also be.  This assumption was not verified
before trying to work with those types, so when these assumptions were
incorrect, the user would get an internal error instead of a helpful error
message.  This change locks the expected behavior in with error messages for
now.