Further improvements to the docs/ version button
Once we bumped the version number on master from 1.18 to 1.19, it
pointed out something I'd missed before -- that the version number for
1.18 was coded into the menu choices itself.  Here, I'm replacing that
hardcoding with a reference to the version variables and renaming the
variables to "nextRelease" and "stagedRelease".  So at a time like the
present,

* currentRelease would be 1.17
* stagedRelease would be 1.18 (this was formerly 'nextRelease')
* nextRelease would be 1.19 (this was formerly 'nextNextRelease')

Once we release:

* currentRelease will be 1.18
* nextRelease will be 1.19
* stagedRelease can remain 1.19

Then once we have a temporary docs branch for 1.19 at the time of its
release, we'd bump nextRelease to 1.20 (since that's what master would
represent).