Update Dockerfiles for Chapel Release 1.18.0: Release Branch Only (fix)
The previous commit neglected to change the "FROM" reference for the gasnet image.

We build and maintain Docker images on GitHub for every Chapel release:
* `chapel/chapel` (built from a recent debian image with source from the latest
  Chapel release)
  Dockerfile = `util/dockerfiles/Dockerfile`
* `chapel/chapel-gasnet` (the above image, with Chapel rebuilt for COMM=gasnet)
  Dockerfile = `util/dockerfiles/gasnet/Dockerfile`
* `chapel/chapel-master` does not exist on DockerHub (at least not yet).
  If we did start building Docker images from master branch and uploading them
  to DockerHub every night, that is probably the name of the image we would use.

The gasnet Dockerfile should show that the gasnet image was built starting from the
"chapel/chapel" image. As noted, `chapel/chapel-master` does not exist on Docker Hub.