Switch bale codes over to setting CHPL_RT_NUM_THREADS_PER_LOCALE
Previously, there was a numTasksPerLocale config const that was used, but this
only impacted the problem size, not the number of tasks being used for compute.
This was fine for correctness, but would cause issues if for instance we used
that to try and do performance studies with less tasks per node. Just switch
over to CHPL_RT_NUM_THREADS_PER_LOCALE since that will get the number of cores
right in the tasking layer too.