Update knucleotide to work with 1.18
This updates our version of knucleotide to work with 1.18.  The main
difference is that where before a loop expression over an associative
domain was resulting in a 1D array result, now shape preservation is
causing it to be associative, requiring sorting to behave differently
(as Vass has already implemented on master).  While here, I also
brought down the removal of the parSafe argument which we retired
from master some time ago but never submitted to the site (figuring
it was too minor to bother with).