Restore comm=none performance lost in PR 11197.
PR 11197 created common code for strided bulk transfer and changed
comm=none and comm=ugni to call that.  Unfortunately that caused a
performance regression with comm=none for NPB sizes A and B.  In the old
code the memmove() that actually moved the bytes was visible to the
compiler when it was compiling its caller and thus effectively, could be
inlined.  But in the new code that was not possible.  Here, move the
common code into static inline functions whose source is again available
to the compiler at the call sites, restoring the old performance.

While here, change chpl_memcpy() to chpl_memmove() in the "nonblocking"
transfer functions for comm=none.  They always should have been such,
since we don't know what the overlap state is for the source and target
arguments to the NB routines.  In addition to causing the performance
regression PR 11197 effectively made the opposite (unsafe) change, but
apparently without breaking any tests.