Increase commonality in runtime communication diagnostics support.
Move the definition of the atomic diags counter struct (the one that has
the counters the comm layer implementations update on the fly) and the
runtime communication diagnostics/verbose flag checking and temporary
disablement into common code.  This "temporary disablement" is what we
use to ignore comms done for the specific purpose of turning on or off
comm diags or verbose comm job-wide.

While here, remove some highly detailed internals-oriented verbose comm
output having to do with strided bulk transfer in comm=gasnet.  As far
as I can tell we don't test this and it has the look of something that
should be done as debug output if it's done at all.