Don't include the original module name when searching use scopes for only/except
Prior to this change, `use M only M` would act as though it were a proper
statement, even though what was found was the defining module itself.  This
would seem to be okay if the symbol was not referenced but would cause problems
when you tried to use it (say, if you had tried to rename the module this way
and then tried to use the new name).

Adds a check to ResolveScope::getFields which removes the current symbol from
the list of matches.  I did this instead of passing the originating symbol along
because it seemed like this same issue would also impact naming the parent
module of the used module (so I could avoid all false positives instead of
just one).  I need to verify this with a test.