Don't allow signed values for environment variables that specify sizes.
The major functional change here is to disallow signed values, that is,
those with initial '-' characters, in environment variables that are
expected to contain sizes.  Thus things like CHPL_RT_MAX_HEAP_SIZE=-1
will now cause a warning instead of ending up specifying a gigantic heap
due to type punning.  This isn't the major thrust of the work to be done
in the associated PR, since obviously it's easy to specify a gigantic
heap using an unsigned value, but it's still a desirable change.

The bulk of the changes here are really a "while I was here", to allow
printing the name of the offending environment variable when we want to
emit a warning about its value.