Do not set certain hugepage-related environment vars with fixed heap.
With a fixed heap, we do not need HUGETLB_NO_RESERVE set to anything.
Some earlier work of mine seems to have resulted in it getting set to
"yes" any time we were using hugepages, even when the heap was of fixed
size.  This broke the code that used hugepage allocation failures to
maximize a fixed heap size, because with HUGETLB_NO_RESERVE=yes even
ridiculously large allocations (for example, more than the physical
memory present) succeed, only to cause failures later when the pages
cannot be created.

Here, only set HUGETLB_NO_RESERVE and CHPL_JE_MALLOC_CONF in the
launcher when the heap is dynamically sized.  Adjust the sanity checking
in the runtime to only complain when the heap is dynamically sized, and
improve the warning messages it emits so that this is more clear.
Adjust the related documentation.  And finally, make the expected output
for the hugepage sanity test match the new messages.