build_configs/cray: chapel-build script needs to forget venv_py27
The previously-developed chapel-build.bash script missed another
requirement for the first round of Jenkins build script replacement:
If building an aarch64 RPM, the build script must not use the
"-B +venv_py27" argument to the setenv script, because the aarch64
setup script does not know that component. It isn't needed, because
any possible build machine for an aarch64 RPM will be at least CLE-6,
with Python 2.7 already installed on the system.

However, when we build for x86_64, we may be building on the old
CLE-5 XC's. The xc_x86_64 setenv script will need to run that
"venv_py27" component. The xe_x86_64 setenv script (which does not
exist yet) will not.

This change simply stops chapel-build from adding "-B +venv_py27"
to the command line when it calls the setenv script.
That's good enough for Cray-XC on aarch64.

Before trying to use this for Cray-XC or XE x86_64 builds, more
changes will be needed here. May just add some kind of if test(s)
to the body of the affected setenv scripts, and do away with that
"-B +venv_py27" command line parameter.