Emit oversized heap warning and if XE, reduce to fit; add a test.
Improve oversized heap messaging.  On XE systems, force the heap size
below the limit, since if it's too big we'll fail when registering it.
Emit a warning saying we've done so.  On XC, just emit a warning about
reduced performance if it's oversized, since it'll actually work in that
case, it just might run slowly.

Add a test for the oversized heap messages.