build_configs/cray: setenv-xc-x86_64 builds venv_py27 automatically if needed
Replaces the command line parameter "-B +venv_py27" with some built-in
if tests in the body of the setenv script, to perform the same function.

This change removes the last bit of locality remaining in the
chapel-build.bash script. All project-specific or machine-dependent
knowledge is contained in the setenv script file, as intended.

Also see the log message from previous commit.