Fix valgrind failure in skyline.chpl
The valgrind failure was caused by a call to this function
when creating the 'result' array in the shapeful chpl__initCopy():

  proc iteratorToArrayType(type t:_iteratorRecord) type {
    var A:[1..0] iteratorToArrayType(__primitive("scalar promotion type", t));
    return A.type;
  }

This function auto-destroys A prior to returning.
That deletes the domain class for A.domain.
However, that class is used as part of the runtime type
that this function returns at run time.
Accessing this runtime type resulted in read-after-delete.

We had not observed this before because we invoked this function
only in the shape-less case where the resulting array DOES have
some elements. In which case the runtime type is unused.
BTW, that happened only in these non-future tests:

  expressions/loop-expr/scoping-1.chpl
  trivial/deitz/statements/nested_fors.chpl
  trivial/deitz/statements/nested_fors2.chpl

The solution is to rewrite the above function to declare A's domain
explicitly and annotate it with pragma "no auto destroy".
This follows the approach in the existing chpl_buildStandInRTT(arrType).

WHILE THERE

* Create that domain using chpl_buildStandInRTT(), for uniformity
with other cases / wider applicability.

* Return chpl__buildArrayRuntimeType() instead of A.type,
as in existing chpl_buildStandInRTT(arrType).

* Replace the separate iteratorToArrayType() with an overload
of chpl_buildStandInRTT() for _iteratorRecord types. That way it can
apply more generally. This overload creates the type of the array
that this iterator creates when stored in a variable.

* Remove "no doc" pragmas because I think this whole file is "no doc".
Remove "inline" annotations in the cases where I do not think they
change anything.

PRIM_STATIC_FIELD_TYPE

I need to get the type of the _shape_ field of the _iteratorRecord
type passed to the new chpl_buildStandInRTT(), formerly iteratorToArrayType().
Because that's the type that we are creating the standin domain for.
For that, I introduce the new primitive, PRIM_STATIC_FIELD_TYPE.
As with PRIM_STATIC_TYPEOF, it returns the compile-time type.
If that time has a runtime component, it will be uninitialized.

Aside reminder: this is the case where result's element type is itself
an (inner) array. We create 'result' before we have seen any
of its upcoming elements, so we cannot use any of the elements'
types. So we are creating the standin domain to use in result's
element type's runtime type.

I used some trickery to avoid introducing this primitive in #10941.
I did not find a way to avoid it here. The closest solutions were:

* Use PRIM_QUERY. That does not work because the _shape_ field
is neither isInstantiatedField(), nor a type or param field.

* Use PRIM_GET_RUNTIME_TYPE_FIELD. That does not seem applicable,
as _iteratorRecord is not a runtime type.

* Create `var ir: irType`. This does not work because _iteratorRecord
types do not have _defaultOf.

Now that we have this primitive, we may want to use it instead of
some of the code from #10941, ex. procs instanceType, domInstanceType,
arrInstanceType.

UPDATE ON test/arrays/shapes/skyline.chpl

It now completes cleanly, and leaks only the standinDomain, as in
the other cases that create a standinDomain. However, each element
of the resulting array is empty. Even though it (each element) IS
assigned a variable-length array. So it remains a .future, just
one more step closer to working.

DETAILS on skyline.chpl FOR THOSE CURIOUS

Cf. test/arrays/ferguson/from-iterator/skyline.chpl
it works properly because the elements of the resulting array
are initialized, not assigned, from the inner arrays
created by makeArray.

Cf. in this code:

  var B1 = [i in {1..n}] makeArray(i);

we get size-mismatch error. Why no such error in our skyline.chpl ?

Because for the statement:

  var B2 = nested forall expression...;

we call:

  B2 = initCopy(ir_outer);

which invokes:

  proc initCopy(ir_outer) {
    var result ....; // array of arrays
    forall (r, src) in zip(result, ir_outer) do
      r = src;
    return result;
  }

where r is an inner array; src is an inner _iteratorRecord.

The assignment "r = src" invokes proc =(ref a: [], b), which invokes

  proc chpl__transferArray(a, b) {
    forall (aa,bb) in zip(a,b) do
      aa = bb;
  }

There, 'a' is the inner array and 'b' is the inner _iteratorRecord.

Because of how we create result's element type, all result's elements -
inner arrays - start out with an empty domain.
The 'forall' runs the DRdom leader on such an inner array.
It either does not yield anything, or yields an empty set of indices.
The corresponding follower of the inner _iteratorRecord
is either not invoked, or is simply not equipped to detect the size mismatch.

Therefore no size-mismatch error is reported.
Also because of this, the inner _iteratorRecord's elements are never
transferred the inner array, so the elements of 'result' remain empty.

Cf. with B1 above, 'src' and 'b' are an inner array. So "r = src"
invokes the array-to-array assignment. That DOES check size mismatch
and does result in the mismatch error at run time.

With contributions by @mppf.

Resolves #11149.