Correct scope resolution of LoopExprs
BACKGROUND

A LoopExpr keeps DefExprs of its index variables in its 'defIndices'.
These index variables are referenced, via SymExprs, from the loop body
as well as in LoopExpr's 'indices'.

We used to revert these SymExprs back to UnresolvedSymExprs at the beginning
of lowering a LoopExpr, because the compiler did not handle them well.
We relied on destructureIndices() and subsequent (additional) scope resolution
to re-create the DefExprs and the SymExprs for the index variables.

This reverting caused trouble because the UnresolvedSymExprs in
LoopExpr::indices did not see the DefExprs created in destructureIndices()
and so used to be scope-resolved incorrectly or not at all.

THIS EFFORT

Now there is no reverting, so the correct SymExprs for the index variables
remain in tact. The rest of the compiler is updated to deal with this.

The key change is that now there can be DefExprs inside non-statement Exprs.
Therefore scope resolution needs to record these DefExprs properly in
'sScopeMap' by calling 'scopeResolve()' functions on things within these Exprs.
For that, this PR adds a few calls to 'scopeResolveExpr()' to scopeResolve.cpp.

At the same time, lowering of LoopExprs in LoopExpr.cpp needed an adjustment.
Lowering of a given LoopExpr creates several copies of the loop body and the
'indices' list. Each copy needs its own Symbols for index variables. So
this PR adds code to clone index variables' DefExprs in each case and to
use the resulting SymbolMap when cloning the 'indices' list and the loop body.
Finally, the original index variables' DefExprs need to be moved from
their LoopExpr node - as it is being eliminated from the AST - into the
appropriate BlockStmt in the newly-created ForLoop.

While there, slightly adjust #include directives to remove ResolveScope.cpp's
dependence on ForallStmt.h.