Add two tests to lock in behavior for exporting functions with array returns
Specifically covers declaring the return but omitting the element type (both
with and without an explicit domain).

Though it would be preferable for the element type to be explicit so that
clients of the library can rely upon a specific type for the returned array,
the library developer is not forced into declaring it.

This behavior already occurred, but was not tracked by any test, so these tests
are adding to ensure it continues to work going forward.