Add error message for Python compilation when array ret element type is omitted
Prior to this change, the error message would be generated by Cython and so
would be a bit cryptic.  This makes the error message clearer until we add
support for this functionality.

I added the error message to the spot where we make the wrapper for exported
functions involving arrays, since that was where it was easiest to distinguish
between explicitly returning chpl_external_array and the return type I described