Add chpl_task_getFixedNumThreads(), and one implementation thereof.
Add a new tasking layer function that, if the implementation runs tasks
by hosting them on a fixed number of threads, returns the number of such
threads.  Otherwise it returns 0 (zero).  So far only tasks=qthreads
returns a non-zero value for this function.  tasks=massivethreads would
do so as well, I believe, but for now I've left that as future work.
For fifo this function returns 0.

This will be used in the comm=ofi implementation, and we have discussed
using it in the comm=ugni implementation.