Get comm=ofi initialization working.
Most of the changes here involve a wholesale replacement of the
previously existing init code with stuff brought over more or less
directly from the ofi mockup.  There are some changes due to the
environment the mockup code finds itself now running in, for example the
provider name now comes from an environment variable and the number of
transmit contexts now comes (preferably) from the tasking layer thread
count, where both of those came from the command line in the mockup.
The dynamic memory allocation interface is different now, too: we use
the Chapel runtime memory management interface instead of the malloc()
interface.  The logic flow is largely unchanged, however.

I also converted the mockup's VERBOSE interface for expected output into
the comm layer's DBG interface for debug-only output.

There are a few TODOs in here for things I haven't decided yet but that
aren't pressing and/or don't seem to have functional implications.
There are also a number of places where variables and functions are
hidden using the preprocessor.  In most cases these are just to avoid
"defined but not used" compiler errors, for things we won't reference
until the AM and/or RDMA code arrives.  (For now, the AM and RDMA
functions just produce internal errors.)

While I was here, I also brought back the comm diagnostics functions.
These had been in a file of their own, up until a few commits ago on
this same branch.  I deleted that file as part of a general reduction in
the number of source files in this comm layer, but when I did so I
failed to move the code itself into the main file.  I only discovered
that mistake when I finally got to the point where I could link user
programs, at which point I got undefs for the comm diags functions.

On 2 nodes of a Cray XC system, hello.chpl built with this gets through
initialization with the sockets provider and then halts as expected the
first time it tries to communicate.  With the gni provider it gets
through initialization up to the memory registration step, where it
fails because there isn't anything to register yet, also as expected.
The shutdown code isn't tested, since the test program always takes the
error exit path, which doesn't invoke it.  I'm not concerned about that,
though, because shutdown is much simpler than setup and if there are
problems in shutdown when we get that far, they should be easy to fix.