Cleanup upon an error in initCopy(_iteratorRecord)
When creating an array using a loop expression and the computation
of a to-be array element throws an exception, do not create the array.
Instead, clean up the to-be array elements that have already been computed.

This came from observing the bug in #11301, where initCopy(_iteratorRecord)
is the function in which pragma "init copy fn" interferes.
I added the pragma "unchecked throws" so that client code does not have
to insert try or throws annotations.

I needed to adjust .good due to #11300. I added a .readme to explain
that the .good is not the ultimately-desired output.

This removes memory leaks in the tests loopexprs-caught, loopexprs-uncaught.

While there, removed a stray comment, as suggested by DavidI.

Future work:
* Add similar cleanup to the shape-ful version of initCopy(_iteratorRecord).
* Split the tests loopexprs-caught, loopexprs-uncaught so that all three
  variants on creating an array using a loop expression get tested.
* Test the case where it is the iterator that throws, not the computation
  of the current element.