Simplify comm substrate logic.
Only comm=gasnet cares about the communication substrate at this point,
so in the tidiness check for the substrate setting, just check for
comm set to gasnet instead of checking for comm not set to any of the
others.  This is also the same thing we do for the segment, immediately
following.