Add support for a fixed heap on hugepages.
This adds comm=ofi versions of chpl_comm_impl_regMemHeapInfo() and
chpl_comm_impl_regMemHeapPageSize(), supporting a fixed heap on
hugepages when using the gni provider.  (That provider's support for
scalable memory registration is limited, so it will be much less work
for us to require a fixed heap when it is used.)

This is mostly all adapted from the comm=ugni implementation.

So far the non-hugepage case is not covered, and the heap is not
registered either way.  Those are next.  However, I've confirmed from
debug output that when we do have hugepages we're getting a heap of the
desired size allocated from the hugepage library.

As a necessary side effect, this also introduces a tiny interface to
libhugetlbfs, so that we can link user programs with that when we have
it and avoid undefined symbols when we don't.