Register the hugepage-based fixed heap.
Get rid of the support for multiple memory region sets, which we
inherited from the ofi mockup and were an experimental feature even
there.  Also get rid of the statically-defined memory regions we
inherited from the mockup.  Replace the latter with dynamically added
entries in a table of memory regions, which only needs to hold 10
entries (for now) since we don't need to support dynamic registration.
(Non-gni providers have scalable registration and will cover all of
memory with a single registration, and for the gni provider we'll just
register the fixed heap.)  Do the registrations.  Confirm they are
working by means of debug output.

hello.chpl now runs on 2 Cray XC nodes with either the sockets or gni
providers and reaches the point where it gets an internal error when
trying to do its first communication.  In other words, we've gotten
through registration successfully with the gni provider.