Rename 4 runtime files.
Rename 4 runtime source files, as follows:
  include/chpl-align.h -> include/chpl-util.h
  include/chpl_rt_utils_static.h -> include/chpl-tasktable.h
  include/chplexit.h -> include/chpl-exit.h
  src/chplexit.c -> src/chpl-exit.c

I'm renaming chpl-align because I'm going to add a couple of new utility
routines which might be useful in any runtime layer, and we don't have a
great vanilla "utility" module yet (surprisingly).  So, I'm co-opting
chpl-align to form chpl-util.  I'm renaming chpl_rt_utils_static because
its name seems to imply that it is a general utility module but in fact
it contains the extern definitions for some task table functions which
are currently in the module code.  I'm renaming chplexit.c because I'm
going to change it anyway since it includes chpl_rt_utils_static when it
doesn't need to and it's always bothered me that chpl-init has a dash
and chplexit does not.  And finally, I'm renaming chplexit.h so that it
matches (renamed) chpl-exit.c.