Snapshot
I spent the afternoon making "solutions" to the original instructions,
with minor edits, in solutions/c-ray-*.chpl.  Then I hit step 9 and
remembered that the problem with the distributed memory version is not
simply that we need to localize a bunch of variables but that the
routines calculating the ray tracer itself need to take those
localized objects as arguments.  I started into this effort in
c-ray-oop.chpl, but it still needs more work.  And then those changes
will need to be back-propagated to the other routines.

Next Q: Can I make "scene" a single record type and then localize it?
Seems like it but I think I should try before getting things wrong
again.