Add common functions to snprintf() nnn[KMG] values.
comm=ugni had a couple of internal functions for snprintf()ing a value
of type size_t or double, in the form 'nnnG', 'nnnM', or 'nnnK' based on
whether they were >= 2**30, >= 2**20, or otherwise.  Here, publicize
these for common use in chpl-format, adjust them to properly handle
values less than 2**10, and replace use of the originals in the ugni
comm layer.

We will use these in comm=ofi soon also.